pyproject.toml
[project]
name = "lore-hound"
version = "0.1.0"
description = "A tenacious AI agent that hunts through codebases, sniffs out architectural patterns, and leaves a trail of cited findings in its Django-PostgreSQL kennel."
requires-python = ">=3.11"
dependencies = [
    "dj-database-url>=3.1.2",
    "django>=5.2.14",
    "django-cors-headers>=4.9.0",
    "django-stubs>=6.0.4",
    "djangorestframework>=3.17.1",
    "djangorestframework-stubs>=3.17.0",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "psycopg[binary]>=3.2",
    "python-dotenv>=1.2.2",
    "whitenoise>=6.9.0",
    "pytest-django>=4.12.0",
]

[dependency-groups]
dev = [
    "basedpyright>=1.39.4",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "lorehound.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
testpaths = ["research"]
